libnetwork: move resolver tests that were skipped on Windows
